backend/requirements.txt
django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-filter==24.3
Pillow==11.1.0
reportlab==4.2.5
python-dotenv==1.0.1
celery==5.4.0
redis==5.2.1
